inserting high scores first try
